when checking access to SYSCONFDIR, only check if it exists, or if it's outside of $localbase XXX cope with mtree

+++ ports/infrastructure/install/setup.ksh
@@ -1,5 +1,5 @@
 #!/bin/mksh
-# $MirOS: ports/infrastructure/install/setup.ksh,v 1.1.2.19 2005/09/12 20:41:16 tg Exp $
+# $MirOS: ports/infrastructure/install/setup.ksh,v 1.1.2.20 2005/09/12 21:39:43 tg Exp $
 #-
 # Copyright (c) 2005
 #	Thorsten "mirabile" Glaser <tg@66h.42h.de>
@@ -208,11 +208,13 @@ if [[ $x = 0 ]]; then
 	print -u2 "Error: Cannot use UID '$myuid', GID '$mygid'!"
 	exit 1
 fi
-x=0; print -n >$etc/$test && chown $myuid $etc/$test && chgrp $mygid $etc/$test \
-    && chown $myuid:$mygid $etc/$test && rm $etc/$test && x=1
-if [[ $x = 0 ]]; then
-	print -u2 "Error: Cannot access '$etc' for UID '$myuid', GID '$mygid'!"
-	exit 1
+if [[ -d $etc || $etc != @($localbase/)* ]]; then
+	x=0; print -n >$etc/$test && chown $myuid $etc/$test && chgrp $mygid $etc/$test \
+	    && chown $myuid:$mygid $etc/$test && rm $etc/$test && x=1
+	if [[ $x = 0 ]]; then
+		print -u2 "Error: Cannot access '$etc' for UID '$myuid', GID '$mygid'!"
+		exit 1
+	fi
 fi
 [[ $myuid = root ]] || export BINOWN=$myuid
 [[ $mygid = bin ]] || export BINGRP=$mygid
